backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
appnope==0.1.4
asttokens==2.4.1
cachetools==6.2.4
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
dnspython==2.8.0
email-validator==2.3.0
executing==2.1.0
fastapi==0.115.2
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-core==2.5.0
google-cloud-firestore==2.22.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jinja2==3.1.6
jupyter-client==8.6.2
jupyter-core==5.7.2
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.2
pluggy==1.6.0
prompt-toolkit==3.0.47
proto-plus==1.27.0
protobuf==6.33.2
psutil==6.0.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pygments==2.18.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pyyaml==6.0.3
pyzmq==27.1.0
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.47.0
shellingham==1.5.4
six==1.16.0
sqlalchemy==2.0.45
sqlmodel==0.0.27
stack-data==0.6.3
starlette==0.40.0
tornado==6.4.1
traitlets==5.14.3
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.1
uvicorn==0.30.1
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.13
websockets==15.0.1
